+++ src/RoomNew/hooks/useRoomForm.ts
@@ -1,28 +1,20 @@
 import { useMutation } from '@tanstack/react-query';
-import z from 'zod';
-import { SubmitErrorHandler, SubmitHandler, useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
+import { SubmitErrorHandler, SubmitHandler, useForm } from 'react-hook-form';
+import z from 'zod';
 import roomAPI from '@/core/api/functions/roomAPI';
 import {
-  ANNOUNCEMENT,
+  FORM_MESSAGE,
+  ROOM_TYPES,
+  TIME_RANGE,
   ROUTINE_NAME,
   ROOM_NAME,
   USER_COUNT,
-  PASSWORD,
-  FORM_MESSAGE
+  PASSWORD
 } from '@/RoomForm/constants/literals';
 
 export const formSchema = z.object({
-  title: z
-    .string()
-    .trim()
-    .min(ROOM_NAME.min, FORM_MESSAGE.ROOM_NAME)
-    .max(ROOM_NAME.max, FORM_MESSAGE.ROOM_NAME),
-  announcement: z
-    .string()
-    .trim()
-    .min(ANNOUNCEMENT.min, FORM_MESSAGE.ANNOUNCEMENT)
-    .max(ANNOUNCEMENT.max, FORM_MESSAGE.ANNOUNCEMENT),
+  type: z.enum(ROOM_TYPES),
   certifyTime: z.number(),
   routines: z.array(
     z.object({
@@ -33,6 +25,11 @@ export const formSchema = z.object({
         .max(ROUTINE_NAME.max, FORM_MESSAGE.ROUTINE_NAME)
     })
   ),
+  title: z
+    .string()
+    .trim()
+    .min(ROOM_NAME.min, FORM_MESSAGE.ROOM_NAME)
+    .max(ROOM_NAME.max, FORM_MESSAGE.ROOM_NAME),
   userCount: z
     .number()
     .gte(USER_COUNT.min, FORM_MESSAGE.USER_COUNT)
@@ -50,21 +47,20 @@ export const formSchema = z.object({
 
 export type Inputs = z.infer<typeof formSchema>;
 
-interface useFormManagementProps {
-  roomId: number;
-  defaultValues: Inputs;
-}
-
-const useFormManagement = ({
-  roomId,
-  defaultValues
-}: useFormManagementProps) => {
+const useRoomForm = () => {
   const mutation = useMutation({
-    mutationFn: roomAPI.putRoom
+    mutationFn: roomAPI.postRoom
   });
 
   const form = useForm<Inputs>({
-    defaultValues,
+    defaultValues: {
+      type: 'MORNING',
+      certifyTime: TIME_RANGE['MORNING'][0],
+      routines: [{ value: '' }],
+      userCount: 5,
+      title: '',
+      password: ''
+    },
     mode: 'onBlur',
     resolver: zodResolver(formSchema)
   });
@@ -72,43 +68,47 @@ const useFormManagement = ({
   const onSubmit: SubmitHandler<Inputs> = (data) => {
     mutation.mutate(
       {
-        roomId,
         title: data.title,
-        announcement: data.announcement,
-        certifyTime: data.certifyTime % 24,
+        password: data.password,
+        type: data.type,
         routine: data.routines.map((r) => r.value),
-        maxUserCount: data.userCount,
-        password: data.password
+        certifyTime: data.certifyTime % 24,
+        maxUserCount: data.userCount
       },
       {
-        onSuccess: (data) => console.log(data),
-        onError: (error) => {
-          const { setError } = form;
+        onSuccess: (data) => {
+          // TODO: toast({ message: '방이 생성되었습니다.', type: 'success' });
+          console.log(data.message);
 
-          // TODO: 에러 Toast 메시지를 보여줘야 해요.
+          // TODO: 방 상세 페이지로 redirect 해야 해요.
+          console.log('TODO: 방 상세 페이지로 redirect');
+        },
+        onError: (error) => {
+          // TODO: toast({ message: '서버에서 날아온 에러 메시지.', type: 'error' });
           console.log(error.response?.data?.message);
 
           if (error.response?.data?.validation) {
+            const { setError } = form;
+
             const {
               title,
-              announcement,
-              routine,
               password,
+              type,
+              routine,
               certifyTime,
               maxUserCount
             } = error.response.data.validation;
 
             setError('title', { message: title });
-            setError('announcement', { message: announcement });
-            setError('routines', { message: routine });
             setError('password', { message: password });
+            setError('type', { message: type });
+            setError('routines', { message: routine });
             setError('certifyTime', { message: certifyTime });
             setError('userCount', { message: maxUserCount });
           }
 
           if (error.response?.status === 401) {
             // TODO: 로그인 페이지로 redirect 해야 해요.
-            // TODO: 혹은 전역 MutationCache에 리다이렉션 관련 로직을 작업해야 해요.
             console.log('TODO: 로그인 페이지로 redirect');
           }
         }
@@ -123,4 +123,4 @@ const useFormManagement = ({
   return { form, handleSubmit, mutation };
 };
 
-export default useFormManagement;
+export default useRoomForm;
