+++ src/RoomNew/steps/SummaryStep.tsx
@@ -3,7 +3,7 @@ import { useFormContext } from 'react-hook-form';
 import { formatHourString } from '@/TimePicker/utils/hour';
 import { BIRD } from '@/RoomForm/constants/literals';
 import { headingStyle, descriptionStyle } from '../constants/styles';
-import { Inputs } from '../hooks/useFormManagement';
+import { Inputs } from '../hooks/useRoomForm';
 import { Icon, iconMap } from '@/shared/Icon';
 
 interface Info {
