+++ src/RoomNew/steps/TimeStep.tsx
@@ -6,7 +6,7 @@ import {
   descriptionStyle,
   errorStyle
 } from '../constants/styles';
-import { Inputs } from '../hooks/useFormManagement';
+import { Inputs } from '../hooks/useRoomForm';
 import { TimePicker } from '@/TimePicker';
 
 const TimeStep = () => {
